better class structure
