+++ src/models/playerAsset/Assets/Structures/Capital.java
@@ -1,14 +1,11 @@
 package models.playerAsset.Assets.Structures;
 
-import models.playerAsset.Assets.ResourceStructure;
 import models.playerAsset.Assets.Worker;
 
 import java.util.ArrayList;
 
-public class Capital extends Structure implements ResourceStructure{
+public class Capital extends ResourceStructure{
 
-    ArrayList<Worker> gatherers;
-    ArrayList<Worker> producers;
 
     public Capital()
     {
@@ -36,35 +33,4 @@ public class Capital extends Structure implements ResourceStructure{
         return "Capital";
     }
 
-    @Override
-    public ArrayList<Worker> removeWorkersFromGathering(int num) {
-        ArrayList<Worker> newList = new ArrayList<>();
-        for (int i = 0; i < num && i < gatherers.size(); i++){
-            newList.add(gatherers.get(i));
-            gatherers.remove(i);
-        }
-
-        return  newList;
-    }
-
-    @Override
-    public ArrayList<Worker> removeWorkersFromProduction(int num) {
-        ArrayList<Worker> newList = new ArrayList<>();
-        for (int i = 0; i < num && i < producers.size(); i++){
-            newList.add(producers.get(i));
-            producers.remove(i);
-        }
-
-        return newList;
-    }
-
-    @Override
-    public void addWorkersToGathering(ArrayList<Worker> list) {
-        gatherers.addAll(list);
-    }
-
-    @Override
-    public void addWorkersToProduction(ArrayList<Worker> list) {
-        producers.addAll(list);
-    }
 }
