+++ src/models/playerAsset/Assets/Structures/Farm.java
@@ -1,6 +1,5 @@
 package models.playerAsset.Assets.Structures;
 
-import models.playerAsset.Assets.ResourceStructure;
 import models.playerAsset.Assets.Worker;
 
 import java.util.ArrayList;
@@ -8,10 +7,8 @@ import java.util.ArrayList;
 /**
  * Created by Clay on 2/26/2017.
  */
-public class Farm extends Structure implements ResourceStructure{
+public class Farm  extends ResourceStructure{
 
-    ArrayList<Worker> gatherers;
-    ArrayList<Worker> producers;
     public Farm() {
 
         setOffDamage(75);
@@ -27,28 +24,4 @@ public class Farm extends Structure implements ResourceStructure{
 
     }
 
-    public ArrayList<Worker> removeWorkersFromGathering(int num){
-        ArrayList<Worker> newList = new ArrayList<>();
-        for (int i = 0; i < num && i < gatherers.size(); i++){
-            newList.add(gatherers.get(i));
-            gatherers.remove(i);
-        }
-
-        return newList;
-    }
-    public ArrayList<Worker> removeWorkersFromProduction(int num){
-        ArrayList<Worker> newList = new ArrayList<>();
-        for (int i = 0; i < num && i < producers.size(); i++){
-            newList.add(producers.get(i));
-            producers.remove(i);
-        }
-
-        return newList;
-    }
-    public void addWorkersToGathering(ArrayList<Worker> list){
-        gatherers.addAll(list);
-    }
-    public void addWorkersToProduction(ArrayList<Worker> list){
-        producers.addAll(list);
-    }
 }
